+++ examples/lang-err-integration-tests/contract-ref/lib.rs
@@ -17,10 +17,7 @@ mod contract_ref {
                 .endowment(0)
                 .code_hash(flipper_code_hash)
                 .salt_bytes(salt)
-                .instantiate()
-                .unwrap_or_else(|error| {
-                    panic!("failed at instantiating the Flipper contract: {:?}", error)
-                });
+                .instantiate();
 
             Self { flipper }
         }